requirements.txt
neat_python==0.92
pygame==2.6.1
